csb: fix ABI types for movement ticks and thing handles
Use uint16_t for disabled_movement_ticks/projectile_disabled_movement_ticks
(G0310/G0311) and thing handle iteration in F0267, matching the original
16-bit DOS ABI. Fix boot profile smoke test linker by adding missing
fmtowns_portrait source.

+++ include/csb_v1_movement_command_step_runtime_pc34_compat.h
@@ -46,8 +46,8 @@ int csb_v1_movement_command_step_runtime_apply_pc34_compat(
 int csb_v1_movement_command_step_runtime_process_queue_pc34_compat(
     CSB_V1_RuntimeProfile *profile,
     struct Dm1V1InputCommandQueuePc34Compat *queue,
-    int disabled_movement_ticks,
-    int projectile_disabled_movement_ticks,
+    uint16_t disabled_movement_ticks,
+    uint16_t projectile_disabled_movement_ticks,
     int last_projectile_disabled_movement_direction,
     CSB_V1_MovementStepWallProbePc34Compat wall_probe,
     void *wall_probe_context,
